Bugfix: * Fixed possible infinite loop in processDefaultArg * constantsTable needs to be allocated on first run * Fixed default parameter parsing

+++ instruction.c
@@ -52,7 +52,7 @@ void processDefaultArg(Context *context, uint32_t *paramCount)
     if (missingCount > 0 && missingCount <= context->defaultCount) {
         // First argument is first on stack, just append defaults
         // which are in constant table in opposite order (last is first)
-        uint32_t index = context->defaultStart + missingCount - 1;
+        int64_t index = context->defaultStart + missingCount - 1;
         for (; index >= context->defaultStart; index--)
             generateInstruction(IST_PushC, 0, index, 0);
         *paramCount += missingCount;
